+++ WorkingSource/D4POSITI.C
@@ -66,7 +66,7 @@ double S4FUNCTION d4position( DATA4 *data )
             memcpy( info->tagName, tag->tagFile->alias, LEN4TAG_ALIAS  ) ;
             info->tagName[LEN4TAG_ALIAS] = 0 ;
             // AS Oct 24/01 - Need to resolve potential tag aliasing problems (t4indx1) */
-            u4ncpy( info->indexFileName, tag->tagFile->indexFile->accessName, strlen( tag->tagFile->indexFile->accessName ) + 1 ) ;
+            u4ncpy( info->indexFileName, tag->tagFile->indexFile->accessName, (unsigned long) strlen( tag->tagFile->indexFile->accessName ) + 1 ) ;
          }
       #endif
       info->startRecno = htonl5(data->recNum) ;
@@ -183,7 +183,7 @@ int S4FUNCTION d4positionSet( DATA4 *data, const double per )
             memcpy( info->tagName, tag->tagFile->alias, LEN4TAG_ALIAS  ) ;
             info->tagName[LEN4TAG_ALIAS] = 0 ;
             // AS Oct 24/01 - Need to resolve potential tag aliasing problems (t4indx1) */
-            u4ncpy( info->indexFileName, tag->tagFile->indexFile->accessName, strlen( tag->tagFile->indexFile->accessName ) + 1 ) ;
+            u4ncpy( info->indexFileName, tag->tagFile->indexFile->accessName, (unsigned long) strlen( tag->tagFile->indexFile->accessName ) + 1 ) ;
          }
       #endif
